type_reference_in_impl_runtime_apis_call
