+++ core/sr-api-macros/tests/ui/type_reference_in_impl_runtime_apis_call.stderr
@@ -1,0 +1,13 @@
+error[E0053]: method `test` has an incompatible type for trait
+  --> $DIR/type_reference_in_impl_runtime_apis_call.rs:20:17
+   |
+14 |         fn test(data: u64);
+   |                       --- type in trait
+...
+20 |         fn test(data: &u64) {
+   |                       ^^^^ expected u64, found &u64
+   |
+   = note: expected type `fn(u64)`
+              found type `fn(&u64)`
+
+For more information about this error, try `rustc --explain E0053`.
